src/backend/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.2
bokeh==3.6.0
boto3==1.35.37
botocore==1.35.37
category-encoders==2.6.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cramjam==2.8.3
cycler==0.12.1
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.9.0
dask-expr==1.1.14
decorator==5.1.1
deprecation==2.1.0
distributed==2024.9.0
dnspython==2.6.1
duckdb==1.0.0
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.114.1
fastapi-cli==0.0.5
fastjsonschema==2.20.0
fastparquet==2024.5.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
imbalanced-learn==0.12.4
importlib_metadata==8.5.0
ipython==8.28.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
locket==1.0.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nbformat==5.10.4
nest-asyncio==1.6.0
netron==7.9.5
nodeenv==1.9.1
numba==0.60.0
numpy==1.24.3
onnxruntime==1.19.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.12.1
orjson==3.10.7
packaging==24.1
pandas==2.1.0
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prisma==0.15.0
prompt_toolkit==3.0.48
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-mon==2.0.5
pyarrow==17.0.0
pycaret==3.3.2
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyod==2.0.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.20.0
s3transfer==0.10.3
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
shellingham==1.5.4
six==1.16.0
sktime==0.26.0
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.38.2
statsmodels==0.14.0
StrEnum==0.4.15
sympy==1.13.2
tbats==1.1.3
tblib==3.0.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tomlkit==0.13.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
tsdownsample==0.1.3
typer==0.12.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchdog==5.0.2
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
wurlitzer==3.1.1
xxhash==3.5.0
xyzservices==2024.9.0
yellowbrick==1.5
zict==3.0.0
zipp==3.20.2
